producer/requirements.txt
Flask==2.0.3
kafka-python==2.0.2
tweepy==4.6.0